findrepo: make compatible with BSD/Mac OS X
Specify '.' as the directory to search for `find`.
Remove the -r option as it's implicit in BSD xargs,
also on GNU, `:|xargs grep` is a noop, as even though
grep is started with no params, it will read from stdin,
which is empty by definition.

This was tested on Mavericks (darwin 13) and Yosemite (darwin 14)