l: propagate exit status from ls
Previously the exit status from this script was
that of the final sed(1), which is usually success,
even if specified files can't be found etc.

Instead use the exit code from the wrapped ls command.

`set -o pipefail` would be an elegant way to achieve that,
but that's bash/ksh/zsh specific, so we manually setup
file descriptors so that the exit status from ls can
be manually output and captured.