l: fix colors for some file types
We were incorrectly matching *.hushlogin=... entries
in $LS_COLORS when removing the coloring for hardlinks (hl=...).
So we adjust the regex matches of $LS_COLORS more restrictive.

Note we pre adjust $LS_COLORS to avoid having
to deal with word boundaries are they were seen to be non-portable.
Testing those showed:
                        \<      \b      [[:<:]]
  solaris sed           yes     no      no
  solaris xpg4 sed      yes     no      error
  netbsd sed            no      no      yes
  gnu sed               yes     yes     yes

Fixes issue #22