pixelbeat-comments: limit Frontend instances used
To limit potential costs as per:
https://cloud.google.com/appengine/docs/managing-costs
Specifically to keep instances less than 28 hours per day.

Note it's not CPU usage, rather reserved instance hours,
so if there were periodic requests (which there are by spammers),
and occasionally overlapping to cause multiple instances to
be started, we can get on average more than a single instance
being continually allocated.

One can monitor the instance usage on the app engine dashboard.
Note there was a spike of "billed instance estimate" to 17
after I changed this setting, but I'm presuming that's because
I was redeploying the app multiple times which may have created
a new instance for each deployment.